phase-5-claud/services/notification-service/requirements.txt
fastapi==0.115.12
uvicorn[standard]==0.34.0
aiokafka==0.11.0
pydantic==2.10.6
python-dotenv==1.0.1
aiosmtplib==3.0.2
httpx
